Refactor PreferencesStore to utilize default preferences
